refactor: Bank 타입을 Bank와 BankManager 타입으로 역할 및 기능 분리
